fix: Move the logic to the component instead of Model
