+++ packages/cozy-procedures/src/components/Documents.jsx
@@ -2,7 +2,6 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { withRouter } from 'react-router'
 import { Title, translate, Label, Button } from 'cozy-ui/transpiled/react/'
-import { AdministrativeProcedure } from 'cozy-doctypes'
 
 import Topbar from './Topbar'
 import EmptyDocumentHolder from './documents/EmptyDocumentHolder'
@@ -12,11 +11,35 @@ import DocumentHolder from './documents/DocumentHolder'
 import { creditApplicationTemplate } from 'cozy-procedures'
 import DocumentsContainer from '../containers/DocumentsDataForm'
 
+/**
+ * This function is used to populate an array based on the order
+ * from the documentsTemplate and set the linked document from the store
+ * @param {Object} docsFromStore
+ * @param {Object} documentsTemplate
+ */
+export const mergeDocsFromStoreAndTemplate = (
+  docsFromStore,
+  documentsTemplate
+) => {
+  let sorted = {}
+  Object.keys(documentsTemplate)
+    .sort(function(a, b) {
+      return documentsTemplate[a].order - documentsTemplate[b].order
+    })
+    .forEach(key => {
+      sorted[key] = documentsTemplate[key]
+      if (docsFromStore[key] && docsFromStore[key].documents) {
+        sorted[key].documents = docsFromStore[key].documents
+      }
+    })
+
+  return sorted
+}
 class Documents extends React.Component {
   render() {
     const { t, router, data: docsFromStore } = this.props
     const { documents: documentsTemplate } = creditApplicationTemplate
-    const populatedTemplateDocsWithStore = AdministrativeProcedure.mergeDocsFromStoreAndTemplate(
+    const populatedTemplateDocsWithStore = mergeDocsFromStoreAndTemplate(
       docsFromStore,
       documentsTemplate
     )
